[DEV-3724] moving to pscHelper namespace
