+++ src/js/models/v2/awardsV2/BaseIdv.js
@@ -3,7 +3,7 @@
  * Created by Lizzie Salita 12/3/18
  */
 
-import * as naicsHelper from 'helpers/naicsHelper';
+import * as pscHelper from 'helpers/pscHelper';
 import CoreLocation from 'models/v2/CoreLocation';
 import CoreAward from './CoreAward';
 import CoreAwardAgency from './CoreAwardAgency';
